requirements-base.txt
feedparser>=6.0
yfinance>=0.2.30
pandas-ta>=0.3.14b1
groq>=0.4.0
pyyaml>=6.0
pydantic-settings>=2.0.0
tqdm>=4.66.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
tenacity>=8.2.0
slowapi>=0.1.9

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
aiofiles>=23.0.0
python-multipart>=0.0.7

# ORM + Database
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Background tasks
apscheduler>=3.10.0

# Notifications
python-telegram-bot>=21.0
